circleci: use `github-release` for managing releases
The `ghr` tools `-delete` argument deletes a release and its git tag in advance if it exists.
The deletion of the tag is likely the reason the build loop is relaunched after the tag is recreated.
Ideally `-delete` should only delete the github release and not perform any operations on the git repo.

In this PR I have replaced `ghr` with the `github-release` tool which is feature rich and does not
delete tags when a release is deleted.

Closes #161

ref:
 - https://github.com/tcnksm/ghr/blob/master/README.md#options